Refactor: Round 3 Step 2.
* Step 2 구현된 내용을 Step 1 으로 병합.
* CoffeeShop.getPrice(of:) 에서 Error 를 발생 하도록 변경.
* CoffeeShop.getPrice(of:) 를 사용 하는 부분에서 Error 를 확인 하도록 변경.
* CoffeeShop.saleMenuItem(of:) 에 @discardableResult 속성 추가.
